test: move submission form tests to own file and add new date tests
